requirements.txt
action-msgs==2.0.2
action-tutorials-interfaces==0.33.5
action-tutorials-py==0.33.5
actionlib-msgs==5.3.6
ament-cmake-test==2.5.4
ament-copyright==0.17.2
ament-cppcheck==0.17.2
ament-cpplint==0.17.2
ament-flake8==0.17.2
ament-index-python==1.8.1
ament-lint==0.17.2
ament-lint-cmake==0.17.2
ament-package==0.16.4
ament-pep257==0.17.2
ament-uncrustify==0.17.2
ament-xmllint==0.17.2
angles==1.16.0
annotated-types==0.7.0
anyio==4.10.0
awscli==1.42.31
botocore==1.40.31
builtin-interfaces==2.0.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
composition-interfaces==2.0.2
cv-bridge==4.1.0
demo-nodes-py==0.33.5
diagnostic-msgs==5.3.6
docutils==0.19
domain-coordinator==0.12.0
example-interfaces==0.12.0
examples-rclpy-executors==0.19.5
examples-rclpy-minimal-action-client==0.19.5
examples-rclpy-minimal-action-server==0.19.5
examples-rclpy-minimal-client==0.19.5
examples-rclpy-minimal-publisher==0.19.5
examples-rclpy-minimal-service==0.19.5
examples-rclpy-minimal-subscriber==0.19.5
fastapi==0.116.1
geometry-msgs==5.3.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
image-geometry==4.1.0
interactive-markers==2.5.4
jmespath==1.0.1
laser-geometry==2.7.1
launch==3.4.5
launch-ros==0.26.8
launch-testing==3.4.5
launch-testing-ros==0.26.8
launch-xml==3.4.5
launch-yaml==3.4.5
lifecycle-msgs==2.0.2
logging-demo==0.33.5
map-msgs==2.4.1
message-filters==4.11.6
mnemonic==0.21
nav-msgs==5.3.6
osrf-pycommon==2.1.6
pcl-msgs==1.0.0
pendulum-msgs==0.33.5
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyNaCl==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-qt-binding==2.2.1
PyYAML==6.0.2
qt-dotgraph==2.7.5
qt-gui==2.7.5
qt-gui-cpp==2.7.5
qt-gui-py-common==2.7.5
quality-of-service-demo-py==0.33.5
rcl-interfaces==2.0.2
rclpy==7.1.4
rcutils==6.7.2
redis==6.4.0
requests==2.32.5
requests-sse==0.5.2
resource-retriever==3.4.4
rmw-dds-common==3.1.0
ros2action==0.32.4
ros2bag==0.26.7
ros2bag-mcap-cli==0.26.7
ros2bag-sqlite3-cli==0.26.7
ros2cli==0.32.4
ros2component==0.32.4
ros2doctor==0.32.4
ros2interface==0.32.4
ros2launch==0.26.8
ros2lifecycle==0.32.4
ros2multicast==0.32.4
ros2node==0.32.4
ros2param==0.32.4
ros2pkg==0.32.4
ros2run==0.32.4
ros2service==0.32.4
ros2topic==0.32.4
rosbag2-interfaces==0.26.7
rosbag2-py==0.26.7
rosgraph-msgs==2.0.2
rosidl-adapter==4.6.5
rosidl-cli==4.6.5
rosidl-cmake==4.6.5
rosidl-generator-c==4.6.5
rosidl-generator-cpp==4.6.5
rosidl-generator-py==0.22.1
rosidl-generator-type-description==4.6.5
rosidl-parser==4.6.5
rosidl-pycommon==4.6.5
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.1
rosidl-typesupport-fastrtps-cpp==3.6.1
rosidl-typesupport-introspection-c==4.6.5
rosidl-typesupport-introspection-cpp==4.6.5
rpyutils==0.4.1
rqt-action==2.2.0
rqt-bag==1.5.4
rqt-bag-plugins==1.5.4
rqt-console==2.2.1
rqt-graph==1.5.4
rqt-gui==1.6.0
rqt-gui-py==1.6.0
rqt-msg==1.5.1
rqt-plot==1.4.3
rqt-publisher==1.7.2
rqt-py-common==1.6.0
rqt-py-console==1.2.2
rqt-reconfigure==1.6.2
rqt-service-caller==1.2.1
rqt-shell==1.2.2
rqt-srv==1.2.2
rqt-topic==1.7.3
rsa==4.7.2
s3transfer==0.14.0
sensor-msgs==5.3.6
sensor-msgs-py==5.3.6
service-msgs==2.0.2
setuptools==80.9.0
shape-msgs==5.3.6
six==1.17.0
sniffio==1.3.1
soroban==0.9.1
sros2==0.13.3
starlette==0.47.3
statistics-msgs==2.0.2
std-msgs==5.3.6
std-srvs==5.3.6
stellar-sdk==13.0.0
stereo-msgs==5.3.6
teleop-twist-keyboard==2.4.0
tf2-geometry-msgs==0.36.12
tf2-kdl==0.36.12
tf2-msgs==0.36.12
tf2-py==0.36.12
tf2-ros-py==0.36.12
tf2-sensor-msgs==0.36.12
tf2-tools==0.36.12
tomli_w==1.2.0
topic-monitor==0.33.5
trajectory-msgs==5.3.6
turtlesim==1.8.3
type-description-interfaces==2.0.2
typing-inspection==0.4.1
typing_extensions==4.15.0
unique-identifier-msgs==2.5.0
urllib3==2.5.0
uvicorn==0.35.0
visualization-msgs==5.3.6
xdrlib3==0.1.1
psycopg2-binary==2.9.9
bcrypt==4.1.2
email-validator==2.1.0
